Set dtype of pd.Index
Fixes FutureWarning: In a future version, the Index constructor will
not infer numeric dtypes when passed object-dtype sequences
(matching Series behavior)

+++ sksurv/compare.py
@@ -112,7 +112,7 @@ def compare_survival(y, group_indicator, return_stats=False):
         table["expected"] = expected
         table["statistic"] = observed - expected
         table = pandas.DataFrame.from_dict(table)
-        table.index = pandas.Index(groups, name="group")
+        table.index = pandas.Index(groups, name="group", dtype=groups.dtype)
         return chisq, pval, table, covar
 
     return chisq, pval
